split batch_inference from inference
